refactor: split tasks into multiple files
